pyproject.toml
[project]
name = "padi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Nealiyer88", email = "neal.iyer88@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=23.2,<24.0",
    "asyncio-throttle>=1.0,<2.0",
    "click>=8.1,<9.0",
    "numpy>=1.26,<2.0",
    "opencv-python-headless>=4.10,<5.0",
    "pathlib2>=2.3,<3.0",
    "pdf2image>=1.17,<2.0",
    "pillow>=10.3,<11.0",
    "psutil>=5.9,<6.0",
    "pydantic>=2.7,<3.0",
    "pymupdf>=1.24,<2.0",
    "pytesseract>=0.3.13,<0.4",
    "python-multipart>=0.0,<1.0",
    "pyyaml>=6.0.1,<7.0",
    "rich>=13.7,<14.0",
    "structlog>=24.1,<25.0",
    "tenacity>=9.0,<10.0",
    "tqdm>=4.66,<5.0",
    "typer>=0.12,<0.13",
    "watchdog>=3.0,<4.0",
]

[project.scripts]
padi = "padi:main"

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"
